backend/requirements.txt
# FastAPI and related dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
alembic>=1.13.1

# Testing
pytest>=7.4.0
httpx>=0.25.0
